Extract type configuration helpers into separate spec component
