+++ spec/features/types/form_configuration_spec.rb
@@ -30,11 +30,11 @@ require 'spec_helper'
 require 'features/projects/project_settings_page'
 
 describe 'form configuration', type: :feature, js: true do
-  let(:admin) { FactoryGirl.create :admin }
-  let(:type) { FactoryGirl.create :type }
+  let(:admin) {FactoryGirl.create :admin}
+  let(:type) {FactoryGirl.create :type}
 
-  let(:project) { FactoryGirl.create :project, types: [type] }
-  let(:category) { FactoryGirl.create :category, project: project }
+  let(:project) {FactoryGirl.create :project, types: [type]}
+  let(:category) {FactoryGirl.create :category, project: project}
   let(:work_package) do
     FactoryGirl.create :work_package,
                        project: project,
@@ -43,125 +43,8 @@ describe 'form configuration', type: :feature, js: true do
                        category: category
   end
 
-  let(:wp_page) { Pages::FullWorkPackage.new(work_package) }
-
-  let(:add_button_dropdown) { page.find '.form-configuration--add-group', text: I18n.t(:label_group) }
-  let(:add_attribute_group_button) { page.find 'a', text: I18n.t('types.edit.add_group') }
-  let(:add_subelements_button) { page.find 'a', text: I18n.t('types.edit.add_subelements') }
-  let(:reset_button) { page.find '.form-configuration--reset' }
-  let(:inactive_group) { page.find '#type-form-conf-inactive-group' }
-  let(:inactive_drop) { page.find '#type-form-conf-inactive-group .attributes' }
-
-  def find_group(name)
-    head = page.find('.group-head', text: name.upcase)
-
-    # Return the parent of the group-head
-    head.find(:xpath, '..')
-  end
-
-  def checkbox_selector(attribute)
-    ".type-form-conf-attribute[data-key='#{attribute}'] .attribute-visibility input"
-  end
-
-  def attribute_selector(attribute)
-    ".type-form-conf-attribute[data-key='#{attribute}']"
-  end
-
-  def find_group_handle(label)
-    find_group(label).find(".group-handle")
-  end
-
-  def find_attribute_handle(attribute)
-    page.find("#{attribute_selector(attribute)} .attribute-handle")
-  end
-
-  def expect_attribute(key:, translation: nil)
-    attribute = page.find(attribute_selector(key))
-
-    unless translation.nil?
-      expect(attribute).to have_selector('.attribute-name', text: translation)
-    end
-  end
-
-  def move_to(attribute, group_label)
-    handle = find_attribute_handle(attribute)
-    group = find_group(group_label)
-    drag_and_drop(handle, group)
-    expect_group(group_label, group_label, key: attribute)
-  end
-
-  def remove_attribute(attribute)
-    attribute = page.find(attribute_selector(attribute))
-    attribute.find('.delete-attribute').click
-  end
-
-  def drag_and_drop(handle, group)
-    target = group.find('.attributes')
-
-    scroll_to_element(group)
-    page
-      .driver
-      .browser
-      .action
-      .move_to(handle.native)
-      .click_and_hold(handle.native)
-      .perform
-
-    scroll_to_element(group)
-    page
-      .driver
-      .browser
-      .action
-      .move_to(target.native)
-      .release
-      .perform
-  end
-
-  def add_attribute_group(name, expect: true)
-    add_button_dropdown.click
-    add_attribute_group_button.click
-
-    input = find('.group-edit-in-place--input')
-    input.set(name)
-    input.send_keys(:return)
-
-    expect_group(name, name) if expect
-  end
-
-  def save_changes
-    # Save configuration
-    # click_button doesn't seem to work when the button is out of view!?
-    scroll_to_and_click find('.form-configuration--save')
-  end
-
-  def rename_group(from, to)
-    find('.group-edit-handler', text: from.upcase).click
-
-    input = find('.group-edit-in-place--input')
-    input.click
-    input.set(to)
-    input.send_keys(:return)
-
-    expect(page).to have_selector('.group-edit-handler', text: to.upcase)
-  end
-
-  def expect_no_attribute(attribute, group)
-    expect(find_group(group)).not_to have_selector("#{attribute_selector(attribute)}")
-  end
-
-  def expect_group(label, translation, *attributes)
-    expect(find_group(translation)).to have_selector(".group-edit-handler", text: translation.upcase)
-
-    within find_group(translation) do
-      attributes.each do |attribute|
-        expect_attribute(attribute)
-      end
-    end
-  end
-
-  def expect_inactive(attribute)
-    expect(inactive_drop).to have_selector(".type-form-conf-attribute[data-key='#{attribute}']")
-  end
+  let(:wp_page) {Pages::FullWorkPackage.new(work_package)}
+  let(:form) {::Components::Admin::TypeConfigurationForm.new}
 
   describe "with EE token" do
     before do
@@ -169,18 +52,18 @@ describe 'form configuration', type: :feature, js: true do
     end
 
     describe 'default configuration' do
-      let(:dialog) { ::NgConfirmationDialog.new }
+      let(:dialog) {::NgConfirmationDialog.new}
       before do
         login_as(admin)
         visit edit_type_tab_path(id: type.id, tab: "form_configuration")
       end
 
       it 'resets the form properly after changes' do
-        rename_group('Details', 'Whatever')
-        expect_attribute(key: :assignee)
+        form.rename_group('Details', 'Whatever')
+        form.expect_attribute(key: :assignee)
 
         # Reset and cancel
-        reset_button.click
+        form.reset_button.click
         dialog.expect_open
         dialog.cancel
 
@@ -192,7 +75,7 @@ describe 'form configuration', type: :feature, js: true do
         sleep 1
 
         # Reset and confirm
-        reset_button.click
+        form.reset_button.click
         dialog.expect_open
         dialog.confirm
 
@@ -200,15 +83,15 @@ describe 'form configuration', type: :feature, js: true do
         sleep 1
 
         expect(page).to have_no_selector('.group-head', text: 'WHATEVER')
-        expect_group('details', 'Details')
-        expect_attribute(key: :assignee)
+        form.expect_group('details', 'Details')
+        form.expect_attribute(key: :assignee)
       end
 
       it 'detects errors for duplicate group names' do
-        add_attribute_group('New Group')
-        add_attribute_group('New Group', expect: false) # would fail since two selectors exist now
+        form.add_attribute_group('New Group')
+        form.add_attribute_group('New Group', expect: false) # would fail since two selectors exist now
 
-        save_changes
+        form.save_changes
 
         expect(page).to have_selector(".type-form-conf-group.-error", count: 1)
       end
@@ -217,34 +100,34 @@ describe 'form configuration', type: :feature, js: true do
         #
         # Test default set of groups
         #
-        expect_group 'people',
-                     'People',
-                     key: :responsible, translation: 'Responsible'
+        form.expect_group 'people',
+                          'People',
+                          key: :responsible, translation: 'Responsible'
 
-        expect_group 'estimates_and_time',
-                     'Estimates and time',
-                     { key: :estimated_time, translation: 'Estimated time' },
-                     { key: :spent_time, translation: 'Spent time' }
+        form.expect_group 'estimates_and_time',
+                          'Estimates and time',
+                          { key: :estimated_time, translation: 'Estimated time' },
+                          { key: :spent_time, translation: 'Spent time' }
 
-        expect_group 'details',
-                     'Details',
-                     { key: :category, translation: 'Category' },
-                     { key: :date, translation: 'Date' },
-                     { key: :percentage_done, translation: 'Progress (%)' },
-                     { key: :priority, translation: 'Priority' },
-                     { key: :version, translation: 'Version' }
+        form.expect_group 'details',
+                          'Details',
+                          { key: :category, translation: 'Category' },
+                          { key: :date, translation: 'Date' },
+                          { key: :percentage_done, translation: 'Progress (%)' },
+                          { key: :priority, translation: 'Priority' },
+                          { key: :version, translation: 'Version' }
 
         #
         # Modify configuration
         #
 
         # Disable version
-        drag_and_drop(find_attribute_handle(:version), inactive_group)
-        expect_inactive(:version)
+        form.drag_and_drop(form.find_attribute_handle(:version), form.inactive_group)
+        form.expect_inactive(:version)
 
         # Rename group
-        rename_group('Details', 'Whatever')
-        rename_group('People', 'Cool Stuff')
+        form.rename_group('Details', 'Whatever')
+        form.rename_group('People', 'Cool Stuff')
 
         # Start renaming, but cancel
         find('.group-edit-handler', text: 'COOL STUFF').click
@@ -255,42 +138,42 @@ describe 'form configuration', type: :feature, js: true do
         expect(page).to have_no_selector('.group-edit-handler', text: 'FOOBAR')
 
         # Create new group
-        add_attribute_group('New Group')
-        move_to(:category, 'New Group')
+        form.add_attribute_group('New Group')
+        form.move_to(:category, 'New Group')
 
         # Delete attribute from group
-        remove_attribute('assignee')
+        form.remove_attribute('assignee')
 
         # Save configuration
-        save_changes
+        form.save_changes
         expect(page).to have_selector('.flash.notice', text: 'Successful update.', wait: 10)
 
         # Expect configuration to be correct now
-        expect_no_attribute('assignee', 'Cool Stuff')
+        form.expect_no_attribute('assignee', 'Cool Stuff')
 
-        expect_group 'Cool Stuff',
-                    'Cool Stuff',
-                    { key: :responsible, translation: 'Responsible' }
+        form.expect_group 'Cool Stuff',
+                     'Cool Stuff',
+                     { key: :responsible, translation: 'Responsible' }
 
-        expect_group 'estimates_and_time',
-                    'Estimates and time',
-                    { key: :estimated_time, translation: 'Estimated time' },
-                    { key: :spent_time, translation: 'Spent time' }
+        form.expect_group 'estimates_and_time',
+                     'Estimates and time',
+                     { key: :estimated_time, translation: 'Estimated time' },
+                     { key: :spent_time, translation: 'Spent time' }
 
-        expect_group 'Whatever',
-                    'Whatever',
-                    { key: :date, translation: 'Date' },
-                    { key: :percentage_done, translation: 'Progress (%)' }
+        form.expect_group 'Whatever',
+                     'Whatever',
+                     { key: :date, translation: 'Date' },
+                     { key: :percentage_done, translation: 'Progress (%)' }
 
-        expect_group 'New Group',
-                    'New Group',
-                    { key: :category, translation: 'Category' }
+        form.expect_group 'New Group',
+                     'New Group',
+                     { key: :category, translation: 'Category' }
 
-        expect_inactive(:version)
+        form.expect_inactive(:version)
 
         # Test the actual type backend
         type.reload
-        expect(type.attribute_groups.map { |el| el.key })
+        expect(type.attribute_groups.map {|el| el.key})
           .to include('Cool Stuff', :estimates_and_time, 'Whatever', 'New Group')
 
         # Visit work package with that type
@@ -337,10 +220,10 @@ describe 'form configuration', type: :feature, js: true do
     end
 
     describe 'required custom field' do
-      let(:custom_fields) { [custom_field] }
-      let(:custom_field) { FactoryGirl.create(:integer_issue_custom_field, is_required: true, name: 'MyNumber') }
-      let(:cf_identifier) { "custom_field_#{custom_field.id}" }
-      let(:cf_identifier_api) { "customField#{custom_field.id}" }
+      let(:custom_fields) {[custom_field]}
+      let(:custom_field) {FactoryGirl.create(:integer_issue_custom_field, is_required: true, name: 'MyNumber')}
+      let(:cf_identifier) {"custom_field_#{custom_field.id}"}
+      let(:cf_identifier_api) {"customField#{custom_field.id}"}
 
       before do
         project
@@ -352,25 +235,25 @@ describe 'form configuration', type: :feature, js: true do
 
       it 'shows the field' do
         # Should be initially disabled
-        expect_inactive(cf_identifier)
+        form.expect_inactive(cf_identifier)
 
         # Add into new group
-        add_attribute_group('New Group')
-        move_to(cf_identifier, 'New Group')
-        expect_attribute(key: cf_identifier)
+        form.add_attribute_group('New Group')
+        form.move_to(cf_identifier, 'New Group')
+        form.expect_attribute(key: cf_identifier)
 
-        page.execute_script('jQuery(".form-configuration--save").click()')
+        form.save_changes
         expect(page).to have_selector('.flash.notice', text: 'Successful update.', wait: 10)
       end
     end
 
     describe 'custom fields' do
-      let(:project_settings_page) { ProjectSettingsPage.new(project) }
+      let(:project_settings_page) {ProjectSettingsPage.new(project)}
 
-      let(:custom_fields) { [custom_field] }
-      let(:custom_field) { FactoryGirl.create(:integer_issue_custom_field, name: 'MyNumber') }
-      let(:cf_identifier) { "custom_field_#{custom_field.id}" }
-      let(:cf_identifier_api) { "customField#{custom_field.id}" }
+      let(:custom_fields) {[custom_field]}
+      let(:custom_field) {FactoryGirl.create(:integer_issue_custom_field, name: 'MyNumber')}
+      let(:cf_identifier) {"custom_field_#{custom_field.id}"}
+      let(:cf_identifier_api) {"customField#{custom_field.id}"}
 
       before do
         project
@@ -380,16 +263,16 @@ describe 'form configuration', type: :feature, js: true do
         visit edit_type_tab_path(id: type.id, tab: "form_configuration")
 
         # Should be initially disabled
-        expect_inactive(cf_identifier)
+        form.expect_inactive(cf_identifier)
 
         # Add into new group
-        add_attribute_group('New Group')
-        move_to(cf_identifier, 'New Group')
+        form.add_attribute_group('New Group')
+        form.move_to(cf_identifier, 'New Group')
 
         # Make visible
-        expect_attribute(key: cf_identifier)
+        form.expect_attribute(key: cf_identifier)
 
-        save_changes
+        form.save_changes
         expect(page).to have_selector('.flash.notice', text: 'Successful update.', wait: 10)
       end
 
@@ -453,7 +336,7 @@ describe 'form configuration', type: :feature, js: true do
   end
 
   describe "without EE token" do
-    let(:dialog) { ::NgConfirmationDialog.new }
+    let(:dialog) {::NgConfirmationDialog.new}
 
     it "should disable adding and renaming groups" do
       with_enterprise_token(nil)
